requirements.txt
# Core Framework
flask==3.0.0
gradio==4.44.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Security
bleach==6.1.0
python-dotenv==1.0.0
cryptography==41.0.7

# Async Support
asyncio-throttle==1.0.2

# Monitoring and Logging
psutil==5.9.6

# Data Validation
pydantic==2.5.0

# Utilities
markupsafe==2.1.3

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For enhanced database support
# psycopg2-binary==2.9.7  # For PostgreSQL
# pymysql==1.1.0  # For MySQL
